tests for paginator merged to default tests
